feat: created test connection and updated runSession

+++ CES_Device_Simulator/Battery.cpp
@@ -5,16 +5,16 @@ Battery::Battery(int n) : powerLevel(n)
 
 }
 
-int Battery::getPowerLevel(){
-    return powerLevel;
+float Battery::getPowerLevel(){ return powerLevel;
 }
 
-void Battery::setPowerLevel(int value){
+void Battery::setPowerLevel(float value){
     powerLevel = value;
 }
 
-int Battery::decrement(int n)
+float Battery::decrement(float n)
 {
     powerLevel -= n;
     return powerLevel;
 }
+
